mlops-pipeline/functions/schema-setup/requirements.txt
functions-framework==3.3.0
google-cloud-bigquery==3.11.4
